requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
authlib==1.6.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.2
colorama==0.4.6
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
fastapi==0.116.1
fastapi-mail==1.5.0
google-auth==2.40.3
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
pillow==11.3.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
qrcode==8.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.42
sqlmodel==0.0.24
starlette==0.47.2
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
